MealItem functionality improve
MealItem component functionality was improved by adding reference with
the help of the useRef hook. Users can now choose a quantity from 1 to 5
for each meal and add it ot their current cart. Quanty cannot be 0 or
empty.

Change was needed for a better UI peroformance an a better UX.